# fixtures/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "BulkRevokeRequest": {
        "properties": {
          "node_id": {
            "default": "",
            "title": "Node Id",
            "type": "string"
          },
          "reason": {
            "default": "bulk revocation",
            "title": "Reason",
            "type": "string"
          },
          "token_ids": {
            "items": {
              "type": "string"
            },
            "title": "Token Ids",
            "type": "array"
          },
          "user_id": {
            "default": "",
            "title": "User Id",
            "type": "string"
          }
        },
        "title": "BulkRevokeRequest",
        "type": "object"
      },
...
      "GrantList": {
        "properties": {
          "grants": {
            "items": {
              "$ref": "#/components/schemas/GrantView"
            },
            "title": "Grants",
            "type": "array"
          }
        },
        "required": [
          "grants"
        ],
        "title": "GrantList",
        "type": "object"
      },
      "GrantView": {
        "properties": {
          "grant_id": {
            "title": "Grant Id",
            "type": "string"
          },
          "node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Id"
          },
          "principal_id": {
            "title": "Principal Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "grant_id",
          "principal_id",
          "role"
        ],
        "title": "GrantView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IssueTokenRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "lifetime_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lifetime Days"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Id"
          },
          "user_id": {
            "default": "",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "IssueTokenRequest",
        "type": "object"
      },
      "IssuedTokenView": {
        "properties": {
          "secret": {
            "title": "Secret",
            "type": "string"
          },
          "token": {
            "$ref": "#/components/schemas/TokenView"
          }
        },
        "required": [
          "token",
          "secret"
        ],
        "title": "IssuedTokenView",
        "type": "object"
      },
...
      "PrincipalView": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "impersonated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impersonated By"
          },
          "impersonating": {
            "default": false,
            "title": "Impersonating",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "principal_id": {
            "title": "Principal Id",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "team_node_id": {
            "default": "",
            "title": "Team Node Id",
            "type": "string"
          }
        },
        "required": [
          "principal_id",
          "kind",
          "display_name",
          "permissions",
          "roles"
        ],
        "title": "PrincipalView",
        "type": "object"
      },
...
      "RevocationResult": {
        "properties": {
          "revoked": {
            "title": "Revoked",
            "type": "integer"
          },
          "token_ids": {
            "items": {
              "type": "string"
            },
            "title": "Token Ids",
            "type": "array"
          }
        },
        "required": [
          "revoked",
          "token_ids"
        ],
        "title": "RevocationResult",
        "type": "object"
      },
...
      "SignInRequest": {
        "properties": {
          "password": {
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "SignInRequest",
        "type": "object"
      },
      "SignInView": {
        "description": "What a successful sign-in hands back: a credential, and when it dies.\n\nThe token is in the body and in nothing else — not a redirect, not a\n``Set-Cookie``. Who stores it and how is the client's decision, and the\nconsole's answer is an HTTP-only cookie the browser cannot read.",
        "properties": {
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "principal_id": {
            "title": "Principal Id",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "expires_at",
          "principal_id"
        ],
        "title": "SignInView",
        "type": "object"
      },
...
      "TokenList": {
        "properties": {
          "tokens": {
            "items": {
              "$ref": "#/components/schemas/TokenView"
            },
            "title": "Tokens",
            "type": "array"
          }
        },
        "required": [
          "tokens"
        ],
        "title": "TokenList",
        "type": "object"
      },
      "TokenView": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "revoked": {
            "title": "Revoked",
            "type": "boolean"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          },
          "team_node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Node Id"
          },
          "token_id": {
            "title": "Token Id",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "token_id",
          "user_id",
          "name",
          "scopes",
          "revoked"
        ],
        "title": "TokenView",
        "type": "object"
      },
...
      "UserList": {
        "properties": {
          "users": {
            "items": {
              "$ref": "#/components/schemas/UserView"
            },
            "title": "Users",
            "type": "array"
          }
        },
        "required": [
          "users"
        ],
        "title": "UserList",
        "type": "object"
      },
      "UserView": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "email",
          "display_name",
          "kind",
          "is_active"
        ],
        "title": "UserView",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Each API version is supported for at least twelve months after the version that replaces it ships. A deprecated version's responses carry a 'Deprecation' header naming the date support ends.",
    "title": "NinjaSRE API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/me": {
      "get": {
        "description": "Return the calling principal and what it may do at its own node (FR-023).",
        "operationId": "whoami_auth_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Whoami",
        "tags": [
          "identity"
        ]
      }
    },
    "/auth/sign-in": {
      "post": {
        "description": "Exchange the local account's name and passphrase for an API token.\n\nPublic by declaration, because it is where the credential every other route\ndemands comes from. A deployment with no local account configured refuses\neverything here, with the same 401 a wrong passphrase gets.",
        "operationId": "sign_in_auth_sign_in_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign In",
        "tags": [
          "identity"
        ]
      }
    },
...
    "/identity/grants": {
      "get": {
        "description": "Return the role grants held in this organisation.",
        "operationId": "list_grants_identity_grants_get",
        "parameters": [
          {
            "in": "query",
            "name": "principal_id",
            "required": false,
            "schema": {
              "default": "",
              "title": "Principal Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Grants",
        "tags": [
          "identity"
        ]
      }
    },
    "/identity/principals": {
      "get": {
        "description": "Return everyone in this organisation.",
        "operationId": "list_principals_identity_principals_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Principals",
        "tags": [
          "identity"
        ]
      }
    },
    "/identity/tokens": {
      "get": {
        "description": "Return this organisation's machine tokens, revoked ones included (FR-026).\n\nRevoked ones are listed rather than filtered out. \"This token was revoked\nlast Tuesday\" is the answer to the question somebody is actually asking when\nthey open this screen.",
        "operationId": "list_tokens_identity_tokens_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "default": "",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tokens",
        "tags": [
          "identity"
        ]
      },
      "post": {
        "description": "Issue a machine token and return its secret exactly once.",
        "operationId": "issue_token_identity_tokens_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuedTokenView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue Token",
        "tags": [
          "identity"
        ]
      }
    },
    "/identity/tokens/revoke": {
      "post": {
        "description": "Revoke a set of tokens at once, selected by id, by owner, or by team.",
        "operationId": "revoke_tokens_identity_tokens_revoke_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkRevokeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevocationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Tokens",
        "tags": [
          "identity"
        ]
      }
    },
    "/identity/tokens/{token_id}": {
      "delete": {
        "description": "Revoke one token.",
        "operationId": "revoke_token_identity_tokens__token_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "token_id",
            "required": true,
            "schema": {
              "title": "Token Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevocationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Token",
        "tags": [
          "identity"
        ]
      }
    },
...
  }
}